[NFC] Use common NumWorkgroups patterns and attributes between LLVM&SPIRV
