+++ iree/compiler/Conversion/CodegenUtils/BUILD
@@ -30,6 +30,7 @@ cc_library(
         "MatmulCodegenStrategy.cpp",
     ],
     hdrs = [
+        "Attributes.h",
         "ForOpCanonicalization.h",
         "FunctionUtils.h",
         "GetNumWorkgroups.h",
