+++ iree/compiler/Conversion/CodegenUtils/CMakeLists.txt
@@ -18,6 +18,7 @@ iree_cc_library(
   NAME
     CodegenUtils
   HDRS
+    "Attributes.h"
     "ForOpCanonicalization.h"
     "FunctionUtils.h"
     "GetNumWorkgroups.h"
