+++ iree/compiler/Conversion/LinalgToSPIRV/Attributes.h
@@ -26,12 +26,6 @@ inline llvm::StringRef getEntryPointScheduleAttrName() {
   return "vkspv.entry_point_schedule";
 }
 
-/// Attribute on a entry point function that specifies which function computes
-/// the number of workgroups.
-inline llvm::StringRef getNumWorkgroupsFnAttrName() {
-  return "vkspv.num_workgroups_fn";
-}
-
 }  // namespace iree_compiler
 }  // namespace mlir
