+++ iree/compiler/Conversion/LinalgToSPIRV/BUILD
@@ -24,9 +24,7 @@ cc_library(
         "ConvertToGPUPass.cpp",
         "ConvertToSPIRVPass.cpp",
         "CooperativeMatrixAnalysis.cpp",
-        "DeclareNumWorkgroupsFnPass.cpp",
         "KernelDispatchUtils.cpp",
-        "LegalizeNumWorkgroupsFnPass.cpp",
         "LinalgTileAndFusePass.cpp",
         "MatMulVectorizationTest.cpp",
         "Passes.cpp",
@@ -45,6 +43,7 @@ cc_library(
     ],
     deps = [
         "//iree/compiler/Conversion/CodegenUtils",
+        "//iree/compiler/Conversion/Common",
         "//iree/compiler/Conversion/HLOToHLO",
         "//iree/compiler/Conversion/HLOToLinalg",
         "//iree/compiler/Conversion/LinalgToVector",
